// src/app/models/master/swuCerriculum.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface SwuCerriculum {
    /** Curriculum SWU ID */
    curriculumSwuId?: number;
    /** Curriculum MHESI ID */
    curriculumMhesiId?: string;
    /** Curriculum Name (Thai) */
    curriculumNameTh?: string;
    /** Curriculum Name (English) */
    curriculumNameEn?: string;
    /** Credit Study */
    creditStudy?: number;
    /** Degree Code */
    degreeCode?: string;
    /** Degree Short Name (Thai) */
    degreeShortTh?: string;
    /** Degree Name (Thai) */
    degreeNameTh?: string;
    /** Degree Short Name (English) */
    degreeShortEn?: string;
    /** Degree Name (English) */
    degreeNameEn?: string;
    /** Major Code */
    majorCode?: string;
    /** Major Name (Thai) */
    majorNameTh?: string;
    /** Major Name (English) */
    majorNameEn?: string;
    /** Education Level Code */
    eduLevelCode?: string;
    /** Education Level Name (Thai) */
    eduLevelNameTh?: string;
    /** Education Level Name (English) */
    eduLevelNameEn?: string;
    /** Study Year */
    studyYear?: number;
    /** Department Code */
    depCode?: string;
    /** Active Flag */
    activeFlag?: boolean;
    /** Create Date */
    createDate?: Date;
    /** Created By ID */
    createById?: number;
    /** Update Date */
    updateDate?: Date;
    /** Updated By ID */
    updateById?: number;
    /** Chairman ID */
    chairmanId?: number;
    /** Owner Department Code */
    ownerDepCode?: string;
    /** Owner Department Name (Thai) */
    ownerDepNameTh?: string;
    /** Owner Department Name (English) */
    ownerDepNameEn?: string;
    /** Owner Faculty Code */
    ownerFacultyCode?: string;
    /** Owner Faculty Name (Thai) */
    ownerFacultyNameTh?: string;
    /** Owner Faculty Name (English) */
    ownerFacultyNameEn?: string;
    depId?: string;
    faculty?: string;
    


    first?: number;
    mode?: string;
    rowNum?: number;
    size?: number;
    filename?: string;
    module?: number;
    prefix?: string;
}