// src/app/models/master/swuSubject.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface SwuSubject {
    subjectSwuId?: number;            // subject_swu_id
    curriculumSwuId?: number;         // curriculum_swu_id
    curriculumMhesiId?: string;       // curriculum_mhesi_id
    groupId?: number;                 // group_id
    groupTh?: string;                 // group_th
    groupEn?: string;                 // group_en
    groupDetail?: string;             // group_detail
    minorTypeId?: number | null;      // minor_type_id
    minorTypeTh?: string | null;      // minor_type_th
    subjectSet?: number;              // subject_set
    subjectCodeTh?: string;           // subject_code_th
    subjectCodeEn?: string;           // subject_code_en
    subjectNameTh?: string;           // subject_name_th
    subjectNameEn?: string;           // subject_name_en
    subjectCredit?: number;           // subject_credit
    subjectTypeTh?: string;           // subject_type_th
    activeFlag?: boolean;             // active_flag
    createDate?: Date;                // create_date
    createById?: number;              // create_by_id
    updateDate?: Date | null;         // update_date
    updateById?: number | null;       // update_by_id
    subjectSwuNo?: number;            // subject_swu_no
    subjectCreditLecture?: number | null;  // subject_credit_lecture
    subjectCreditLab?: number | null;      // subject_credit_lab
    subjectCreditResearch?: number | null; // subject_credit_research
    subjectCreditFormat?: string | null;   // subject_credit_format

    first?: number;
    mode?: string;
    rowNum?: number;
    size?: number;
    filename?: string;
    module?: number;
    prefix?: string;

    groupName?: string;
    minorType?: string;
    subjectCode?: string;
    subjectName?: string;
    subjectType?: string;
}